// packages/components/src/ui/ListItem/ListItem.styles.d.ts
export declare const useStyles: () => {
  item: {
    display: 'flex';
    flexDirection: 'row';
    paddingLeft: import('@alveole/theme').Spacing;
    paddingRight: import('@alveole/theme').Spacing;
    alignItems: 'center';
    gap: import('@alveole/theme').Spacing;
    alignSelf: 'stretch';
  };
  itemHover: {
    backgroundColor: `rgba${string}`;
    borderRadius: import('packages/theme/src/constants/Radius').Radius;
    cursor: 'pointer';
  };
  container: {
    display: 'flex';
    flexDirection: 'row';
    gap: import('@alveole/theme').Spacing;
  };
  detail: {
    paddingTop: import('@alveole/theme').Spacing;
    paddingBottom: import('@alveole/theme').Spacing;
    flex: number;
    display: 'flex';
    flexDirection: 'row';
    alignItems: 'center';
    alignSelf: 'stretch';
    gap: import('@alveole/theme').Spacing;
    position: 'relative';
  };
  separateur: {
    left: number;
    top: number;
    position: 'absolute';
    width: '100%';
    height: number;
    backgroundColor: '#DEE3EC';
  };
  principal: {
    display: 'flex';
    flexDirection: 'column';
    justifyContent: 'center';
    alignItems: 'flex-start';
    flex: number;
  };
  title: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 14;
    lineHeight: 20;
    letterSpacing: 0;
    color: '#373A3F';
  };
  description: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 12;
    lineHeight: 20;
    letterSpacing: 0;
    color: '#5F6571';
  };
  additional: {};
  trailing: {
    display: 'flex';
    justifyContent: 'center';
    alignSelf: 'stretch';
    color: '#5F6571';
  };
  loading: {
    display: 'flex';
    alignItems: 'center';
    justifyContent: 'center';
    color: '#D9127C';
  };
  defaultIcon: {
    color: '#5F6571';
  };
  previewContainer: {
    display: 'flex';
    alignItems: 'center';
    justifyContent: 'center';
    overflow: 'hidden';
    borderRadius: import('@alveole/theme').Spacing;
  };
  preview: {
    width: number;
    height: number;
  };
};
//# sourceMappingURL=ListItem.styles.d.ts.map